requirements.txt
streamlit
llama-index
faiss-cpu
pypdf
llama-index-llms-google-genai
llama-index-embeddings-google-genai
llama-index-embeddings-huggingface
sentence-transformers